Finalizando o cadastro de imoveis

+++ app/Http/Controllers/Admin/CidadeController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
 use App\Cidade;
+use App\Imovel;
 
 class CidadeController extends Controller
 {
@@ -65,6 +66,22 @@ class CidadeController extends Controller
 
     public function deletar($id)
     {
+        //Verficando se possui imoveis relacionados a cidade
+        if (Imovel::where('cidade_id', '=', $id)->count()) {
+
+            $msg = "Não é possível deletar essa cidade!
+            Esses imóveis ( ";
+            $imoveis = Imovel::where('cidade_id', '=', $id)->get();
+            foreach ($imoveis as $imovel) {
+                $msg .= " id: " . $imovel->id . " ";
+            }
+            $msg .= ") estão relacionados.";
+
+            \Session::flash('mensagem', ['msg' => $msg, 'class' => 'red white-text']);
+            return redirect()->route('admin.cidades');
+
+        }
+
         $registro = Cidade::find($id);
 
         $registro->delete();
